feat: implement RunService for managing assistant runs and refactor thread_service
